refactor(dataset): extract bag-download orchestration to its own module
Phase 3 audit §3.A — move the bag-download cluster off ``Dataset``
into a focused module with one external entry point.

WHY
====

``dataset.py`` was a 3400-LoC file mixing two concerns:

  * Dataset domain model — CRUD, versioning (mark_dev / release),
    member management, history, parent/child walking, denormalize
    sugar. The bulk of the file. ~2700 LoC.
  * Bag-download orchestration — five private methods composing
    a three-tier caching strategy (local index, MINID/S3,
    client-side bag generation). ~600 LoC.

The orchestration cluster has exactly **one external entry point**
(``Dataset.download_dataset_bag``) and **one external dependency**
(``DatasetBagBuilder``). It's a self-contained subsystem, and
keeping it in ``dataset.py`` made the Dataset class read like a
laundry list.

WHAT MOVED
==========

New module ``src/deriva_ml/dataset/bag_download.py`` (~650 LoC):

  * Module-level helper:
        _hash_spec(spec)                  (was free function in dataset.py)
  * Free functions (each takes a Dataset arg, used to be methods):
        get_dataset_minid(dataset, ...)
        create_dataset_minid(dataset, ...)
        download_dataset_minid(dataset, ...)
        fetch_minid_metadata(dataset, ...)
        materialize_dataset_bag(dataset, ...)

``Dataset.download_dataset_bag`` stays on ``Dataset`` (it's the
public API users call); its body now calls the free functions from
``bag_download.py``.

WHY FREE FUNCTIONS, NOT A CLASS
================================

The denormalize planner (PR #139) was extracted as a class because
it has state (caching reachability results) and is reused across
multiple denormalization calls on the same model. The bag-download
cluster is stateless — each function takes a dataset and produces a
path/bag. The audit explicitly suggests "bag_pipeline.py" or
"download.py" — module of functions, not class.

OTHER METHODS LEFT ON DATASET
==============================

Some download-adjacent methods stay on ``Dataset`` because they
have different boundaries:

  * estimate_bag_size — separate concern per audit §3.D (the
    async-query loop alignment with bag-pipeline is a separate
    decision point).
  * bag_info / cache — cache-status / cache-warming public API,
    distinct from the download flow.
  * _estimate_csv_bytes / _human_readable_size — helpers for
    estimate_bag_size only.
  * _version_snapshot_catalog / _version_snapshot_catalog_id —
    used by 6+ methods on Dataset, not download-specific.

DEAD-CODE SWEEP
================

  * Removed module-level _hash_spec from dataset.py (only ever
    called from the moved methods; now lives in bag_download.py).
  * Ruff auto-cleaned 20 newly-unused imports from dataset.py
    (hashlib, json, os, shutil, TemporaryDirectory, urlparse,
    hash_utils, bdb, fetch_single_file, DerivaExport, DerivaDownload*
    exception classes, format_exception, requests, DatasetMinid).

NUMBERS
=======

  dataset.py:           3397 LoC → 2814 LoC  (-583)
  bag_download.py:      new      →  650 LoC

  Net: ~50 LoC (mostly the module docstring + function-level
  docstrings about the cluster). The arrangement is clearer and
  the Dataset class is now scoped to "domain model" rather than
  "domain model + bag pipeline".

TEST RESULTS
============

  * tests/local_db/ + tests/asset/ + tests/model/: 434 passed,
    3 skipped, 0 failed.
  * tests/dataset/test_estimate_bag_size.py: 13/13 passed (the
    _estimate_csv_bytes / _human_readable_size mocks still work
    because those methods didn't move).
  * Lint (ruff): clean.
